+++ qgrep/src/qgrep.hpp
@@ -6,4 +6,4 @@
 #define QGREP_DLL __declspec(dllimport)
 #endif
 
-QGREP_DLL void qgrepWrapperAsync(char* arguments, int size, bool (*stringCallback)(const char*, int), void (*errorsCallback)(const char*, int), void (*progressCallback)(int));
+QGREP_DLL void qgrepWrapperAsync(char* arguments, int size, bool (*stringCallback)(const char*, int), void (*errorsCallback)(const char*, int), void (*progressCallback)(double));
