+++ qgrepControls/ToolWindows/SearchWindow/qgrepFilesWindow.xaml.cs
@@ -357,7 +357,7 @@ namespace qgrepControls.SearchWindow
             }
         }
 
-        private void HandleProgress(int percentage)
+        private void HandleProgress(double percentage)
         {
             if (!progressUpdateStopWatch.IsRunning || progressUpdateStopWatch.ElapsedMilliseconds > 20)
             {
